% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allotalot.R
\docType{package}
\name{allotalot-package}
\alias{allotalot}
\alias{allotalot-package}
\title{Parallel Assignment in R}
\description{
allotalot provides tools to perform unpacking and parallel assignment in R.
To get started with allotalot check out the introductory vignette.
}
\seealso{
\code{\link{\%<-\%}}
}

