% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allotalot.R
\docType{package}
\name{allotalot}
\alias{allotalot}
\alias{allotalot-package}
\title{Unpacking Assignment in R}
\description{
allotalot provides an infix function \code{\link{\%<-\%}} to perform
unpacking assignment in R. To get started with allotalot check out the
introductory vignette, \code{browseVignettes('allotalot')}.
}
\seealso{
\code{\link{\%<-\%}}
}

