% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\alias{items}
\title{Enumerate an Object}
\usage{
enumerate(x)

items(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
For \code{items} a list of name, value pairs, one pair for each element of
\code{x}. The name in each pair is the original name of the value in
\code{x}. If element is unnamed, \code{''} is used. See below for examples.

For \code{enumerate} a list of index, value pairs, one pair for each element
of \code{x}. The indices range from 1 to the number of elements in the
unpacked list of \code{x}.
}
\description{
The \code{items} function unpacks a named object into a list of name, value
pairs. The \code{enumerate} function unpacks an object into a list of index,
value pairs.
}
\examples{
# some food groups and choices
foods <- list(
  breads = c('rye', 'wheat'),
  veggies = c('peas', 'spinach', 'corn'),
  fruits = c('plums', 'cherries')
)

# we can use `items` and `enumerate` to
# restructure our list of foods
for (pair in enumerate(items(foods))) {
  {i: {group: choices}}  \%<-\% pair

  cat(paste0(i, '. ', group, '\\n'))
  cat(paste0('    * ', choices, '\\n'), sep = '')
}

# more enumeration, this time unpacking the
# iris data set
for (col in enumerate(iris)) {
  {i: values} \%<-\% col

  if (i != 5) {
    cat('mean', i, 'is', mean(values), '\\n')
  }
}

# alternatively, we could use `items`
# to unpack the iris data set
for (col in items(iris)) {
  {nm: values} \%<-\% col

  if (nm != 'Species') {
    cat(nm, 'mean is', mean(values), '\\n')
  }
}

}
\seealso{
\code{\link{\%<-\%}}
}

