% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massign.R
\name{massign}
\alias{massign}
\title{Assign Values to a Name or Names}
\usage{
massign(x, values, envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{a list of variable name(s), see details.}

\item{values}{values to be assigned, see details.}
}
\description{
Using two lists of names and values, assign values to names in an
environment.
}
\details{
\code{massign} expects \code{x} and \code{values}
}
\keyword{internal}

