% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Parallel, Multiple, and Unpacking Assignment}
\usage{
\%\%(names, values) <- value
}
\arguments{
\item{names}{A bare name or name structure generated using \code{\link{.}}.}

\item{values}{A list of values to assign to names.}
}
\description{
~ missing ~
}
\examples{
.(a, b) \%<-\% list('A', 'B')

print(a)

# swap values
.(a, b) \%<-\% list(b, a)

print(a)
}

