% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{rs_addin_insert_multi_assign}
\alias{rs_addin_insert_multi_assign}
\title{Insert multiple assignment operator}
\usage{
rs_addin_insert_multi_assign()
}
\value{
\code{structure(list(), .Names = character(0))}
}
\description{
Call this function as an addin to insert \code{ \%<-\% } at the
cursor position.
}
\author{
David Kahle \email{david@kahle.io}
}
