% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeallot.R
\docType{package}
\name{zeallot}
\alias{zeallot}
\alias{zeallot-package}
\title{Multiple and Unpacking Assignment in R}
\description{
zeallot provides a \code{\link{\%<-\%}} operator to perform unpacking
assignment in R. To get started with zeallot be sure to read over the
introductory vignette on unpacking assignment,
\code{vignette('unpacking-assignment')}.
}
\seealso{
\code{\link{\%<-\%}}
}

