% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorcher.R
\docType{package}
\name{zorcher}
\alias{zorcher}
\alias{zorcher-package}
\title{Multiple and Unpacking Assignment in R}
\description{
zorcher provides an infix function \code{\link{\%<-\%}} to perform
unpacking assignment in R. To get started with zorcher check out the
introductory vignette, \code{browseVignettes('zorcher')}.
}
\seealso{
\code{\link{\%<-\%}}
}

