% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massign.R
\name{massign}
\alias{massign}
\title{Assign Values to a Name or Names}
\usage{
massign(x, values, envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{a list of variable name(s).}

\item{values}{values to be assigned, usually a list of values or an object
with a \code{destructure} implementation.}
}
\description{
Using two lists of names and values, assign values to names in the specified
environment, defaults to calling environment.
}
\details{
For more information about how \code{massign} and \code{\%<-\%} assign values
to names, refer to the introductory vignette.

\code{browseVignettes('allotalot')}
}
\keyword{internal}

