% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-op.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Parallel, Multiple, and Unpacking Assignment}
\usage{
x \%<-\% value
}
\arguments{
\item{x}{A bare name or name structure.}

\item{value}{A list of values to assign.}
}
\value{
\code{\%<-\%} invisibly returns \code{value}.
}
\description{
The \code{\%<-\%} operator performs parallel assignment by coercing the LHS
of an assignment expression to a name structure, see \code{\link{.}}, and
assigning values from the RHS to these names.
}
\examples{
.(a, b) \%<-\% list('A', 'B')
a  # "A"
b  # "B"

# swap the values
.(a, b) \%<-\% list(b, a)
a  # "B"
b  # "A"

guests \%<-\% c('Nathan,Allison,Matt,Polly', 'Smith,Peterson,Williams,Jones')

.(firsts, lasts) \%<-\% strsplit(guests, ',')

firsts  # "Nathan", "Allison", ..
lasts   # "Smith", "Peterson", ..
}
\seealso{
\code{\link{.}}, \code{\link{unpack}}, \code{\link{items}},
  \code{\link{enumerate}}
}

