% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeallot.R
\docType{package}
\name{zeallot}
\alias{zeallot-package}
\alias{zeallot}
\title{Multiple, unpacking, and destructuring assignment in R}
\description{
zeallot provides a \code{\link{\%<-\%}} operator to perform multiple
assignment in R. To get started with zeallot be sure to read over the
introductory vignette on unpacking assignment,
\code{vignette('unpacking-assignment')}.
}
\seealso{
\code{\link{\%<-\%}}
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nate@haufin.ch}

Other contributors:
\itemize{
  \item Paul Teetor [contributor]
}

}
